+++ src/settingsTab.ts
@@ -3,6 +3,7 @@ import type SourceModeStyling from './main';
 import { addFontFamilySetting } from './settings/FontFamilySetting';
 import { addFontSizeSetting } from './settings/FontSizeSetting';
 import { addLineHeightSetting } from './settings/LineHeightSetting';
+import { addFontColorSetting } from './settings/FontColorSetting';
 import { addHeadingColorSetting } from './settings/HeadingColorSetting';
 import { addBackgroundColorSetting } from './settings/BackgroundColorSetting';
 import { addFontWeightSetting } from './settings/FontWeightSetting';
@@ -28,6 +29,7 @@ export class SourceModeStylingSettingTab extends PluginSettingTab {
 		addFontFamilySetting(containerEl, this.plugin);
 		addFontWeightSetting(containerEl, this.plugin);
 		addFontSizeSetting(containerEl, this.plugin);
+		addFontColorSetting(containerEl, this.plugin);
 		addLineHeightSetting(containerEl, this.plugin);
 		addHeadingColorSetting(containerEl, this.plugin);
 		addBackgroundColorSetting(containerEl, this.plugin);