+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+- Font color setting with theme default and custom color picker
+- Default custom font color set to #333333
+
+### Changed
+- Streamlined essential font list to focus on most reliable detection
+
 ## [0.2.3] - 2025-06-20
 
 ### Fixed